pre-existing peer will be inserted at the end
the current implmentation of InsertPeer() doesn't change the relative
position of pre-existing peer in the peer_list_. So, when a peer already
in the peer_list_ tries to connect  only its losses_ are set to
zero. Now, if an already existing peer has an incoming connection it
will be removed from its current position it the peer_list_ and will be
appended at the end just like an non-existing peer.